% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{gui}
\usage{
gui(envirfile = NULL)
}
\arguments{
\item{envirfile}{A file to a saved environment of a project}
}
\description{
A simple GUI for using dnamatch2
}
\details{
The user can select settings in a GUI for running dnamatch2
}
\author{
Oyvind Bleka <oyvble.at.hotmail.com>
}
